requirements.txt
scikit-learn==0.23.2
scipy==1.5.2
librosa==0.8.0
Pillow==7.1.0
pickleshare==0.7.5
jupyterlab==1.2.6
ipython==7.16.1
torch==1.4.0
torchvision==0.5.0
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.1